backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
openai>=1.58.0
httpx>=0.28.0
python-multipart>=0.0.18
pydantic>=2.10.0
aiofiles>=24.1.0
python-dotenv>=1.0.1
